+++ mg_custom_nodes/KLL535_ComfyUI_Simple_Qwen3-VL-gguf_20260701/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Nightly
+- improvement of video input (part 1)
 - fix UnicodeDecodeError error in subprocess
 - add node `Ideogram 4 JSON Preview` and `Ideogram 4 JSON Swap XY Coordinates`
 - present_penalty/presence_penalty issue